--- internal/filter/BUILD.bazel
@@ -1,14 +1,0 @@
-load("@io_bazel_rules_go//go:def.bzl", "go_library")
-
-go_library(
-    name = "filter",
-    srcs = ["filter.go"],
-    importpath = "github.com/lestrrat-go/jwx/v3/internal/filter",
-    visibility = ["//:__subpackages__"],
-)
-
-alias(
-    name = "go_default_library",
-    actual = ":filter",
-    visibility = ["//:__subpackages__"],
-)
