+++ jws/BUILD.bazel
@@ -27,7 +27,7 @@ go_library(
         "//cert",
         "//internal/base64",
         "//internal/ecutil",
-        "//internal/filter",
+        "//transform",
         "//internal/json",
         "//internal/keyconv",
         "//internal/pool",
